+++ addons/binding/org.openhab.binding.tado/src/main/java/org/openhab/binding/tado/internal/config/TadoHomeConfig.java
@@ -6,16 +6,14 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.tado.config;
+package org.openhab.binding.tado.internal.config;
 
 /**
- * Holder-object for zone configuration
+ * Holder-object for home configuration
  *
  * @author Dennis Frommknecht - Initial contribution
  */
-public class TadoZoneConfig {
-    public long id;
-    public int refreshInterval;
-    public int fallbackTimerDuration;
-    public int hvacChangeDebounce;
+public class TadoHomeConfig {
+    public String username;
+    public String password;
 }
