+++ addons/binding/org.openhab.binding.tado/src/main/java/org/openhab/binding/tado/internal/config/TadoZoneConfig.java
@@ -6,14 +6,16 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.tado.config;
+package org.openhab.binding.tado.internal.config;
 
 /**
- * Holder-object for mobile device configuration
+ * Holder-object for zone configuration
  *
  * @author Dennis Frommknecht - Initial contribution
  */
-public class TadoMobileDeviceConfig {
-    public int id;
+public class TadoZoneConfig {
+    public long id;
     public int refreshInterval;
+    public int fallbackTimerDuration;
+    public int hvacChangeDebounce;
 }
